add exchanges in arbitrage
