+++ app/src/main/res/layout/controls_arbitrage_chart.xml
@@ -5,56 +5,13 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
-    tools:showIn="@layout/fragment_arbitrage">
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginBottom="20dp"
-        android:background="@color/colorPrimaryDark"
-        android:orientation="horizontal">
-
-        <com.jaredrummler.materialspinner.Spinner
-            android:id="@+id/currencyFromSpinner"
-            android:layout_width="100dp"
-            android:layout_height="wrap_content"
-            android:layout_marginLeft="5dp"
-            android:maxLines="1"
-            android:theme="@style/spinnerStyle"
-            android:textColor="@android:color/white"
-            app:ms_arrow_tint="@android:color/white"
-            app:ms_text_color="?android:attr/textColorPrimary"
-            app:ms_dropdown_max_height="300dp" />
-
-        <com.jaredrummler.materialspinner.Spinner
-            android:id="@+id/currencyOneSpinner"
-            android:layout_width="100dp"
-            android:layout_height="wrap_content"
-            android:maxLines="1"
-            android:theme="@style/spinnerStyle"
-            android:textColor="@android:color/white"
-            app:ms_arrow_tint="@android:color/white"
-            app:ms_text_color="?android:attr/textColorPrimary"
-            app:ms_dropdown_max_height="300dp" />
-
-
-        <com.jaredrummler.materialspinner.Spinner
-            android:id="@+id/currencyTwoSpinner"
-            android:layout_width="100dp"
-            android:layout_height="wrap_content"
-            android:layout_marginRight="5dp"
-            android:maxLines="1"
-            android:theme="@style/spinnerStyle"
-            android:textColor="@android:color/white"
-            app:ms_arrow_tint="@android:color/white"
-            app:ms_text_color="?android:attr/textColorPrimary"
-            app:ms_dropdown_max_height="300dp" />
-    </LinearLayout>
+    tools:showIn="@layout/fragment_arbitrage_chart">
 
     <LinearLayout
         android:id="@+id/controls"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
+        android:paddingTop="10dp"
         android:orientation="vertical">
 
         <LinearLayout
